Simplified model code.
After talking with Peter, decided to simplify the code. I took out
the constant resource supply rate (S) and the resource turnover
rate (a) to make resource dynamics simpler. Now the resource (R)
at time t is set before running the model with a mean and sd --
on log scale. This still needs to be test by invasion growth rates
and for ability to simulate relative nonlinearity.